MAILBOX-264 put MailboxMessage repository access into its own class
