+++ mailbox/cassandra/src/test/java/org/apache/james/mailbox/cassandra/CassandraSubscriptionManagerTest.java
@@ -20,11 +20,15 @@
 package org.apache.james.mailbox.cassandra;
 
 import org.apache.james.backends.cassandra.CassandraCluster;
+import org.apache.james.backends.cassandra.init.CassandraTypesProvider;
 import org.apache.james.mailbox.AbstractSubscriptionManagerTest;
 import org.apache.james.mailbox.SubscriptionManager;
+import org.apache.james.mailbox.cassandra.mail.CassandraMessageRepository;
 import org.apache.james.mailbox.cassandra.mail.CassandraModSeqProvider;
 import org.apache.james.mailbox.cassandra.mail.CassandraUidProvider;
 import org.apache.james.mailbox.cassandra.modules.CassandraSubscriptionModule;
+
+import com.datastax.driver.core.Session;
 
 /**
  * Test Cassandra subscription against some general purpose written code.
@@ -35,12 +39,16 @@ public class CassandraSubscriptionManagerTest extends AbstractSubscriptionManage
     
     @Override
     public SubscriptionManager createSubscriptionManager() {
+        Session session = cassandra.getConf();
+        CassandraTypesProvider typesProvider = cassandra.getTypesProvider();
+        CassandraMessageRepository messageRepository = new CassandraMessageRepository(session, typesProvider);
         return new CassandraSubscriptionManager(
             new CassandraMailboxSessionMapperFactory(
-                new CassandraUidProvider(cassandra.getConf()),
-                new CassandraModSeqProvider(cassandra.getConf()),
-                cassandra.getConf(),
-                cassandra.getTypesProvider()
+                new CassandraUidProvider(session),
+                new CassandraModSeqProvider(session),
+                session,
+                typesProvider,
+                messageRepository
             )
         );
     }
